Update evidence index with explicit type mappings
Without explicit type mappings ES infers the type of new document
fields. A small number of evidence objects have values which do not fit
into a `float`. Since the vast bulk of entries do fit into a `float` ES
infers that to be the type. This causes ~50 entries to be discarded due
to a type mismatch.

This bug is non-deterministic: there is always a chance that one of the
50 (out of ~650k) will be loaded in the first batch and the type
   inference will work correctly.

Resolves opentargets/platform#1687