Update odds ratio field for genetics evidence.
The auto-indexing of ES assigned `float` types to these fields. For a
very small number of evidence strings this needs to be `double`.
Presently only the genetics evidence is effected by this change, so have
added in a separate index mapping file with these fields appropriately
changed.

See ticket 1866 for greater detail.

Resolves opentargets/issues#1866
Resolves opentargets/issues#1687