Fix esbulk errors
esbulk seems to have been causing errors with the previous settings.
We did not see them before because they do not halt the image building
process, but the db becomes inconsistent, as documents fail to import.

According to issues in the library github:

https://github.com/miku/esbulk/issues/33

and the readme:

https://github.com/miku/esbulk?tab=readme-ov-file#esbulk

Using default settings might solve those errors. So far, it seems to
be working.